Agrego en el Service y Dao metodo para obtener los peluqueros cercanos y tambien por nombre o tipo dentro del rango de la ubicacion dada, con el objeto pageable para ordenar y/o paginar los resultados.

+++ src/main/kotlin/ar/edu/unq/peluqueriayabackend/persistence/impl/repositories/PeluqueroRepository.kt
@@ -13,6 +13,6 @@ interface PeluqueroRepository : JpaRepository<Peluquero, Int>{
     fun findAllInRangeAtCoordinates(rangoEnKm: Double,lat:Double,lng: Double, pageable: Pageable): Page<Peluquero>
 
     @Query("SELECT p FROM Peluquero p WHERE distance(cast(p.ubicacion.latitude as double),cast(p.ubicacion.longitude as double),?2,?3) <= ?1 AND lower(p.nombre) LIKE lower(concat('%',?4,'%')) OR ?5 member of p.tipos")
-    fun findAllInRangeAtCoordinatesAndLikeNameOrWithTipos(rangoEnKm:Float,lat:Float,lng: Float, nombre: String, tipo: String, pageable: Pageable): Page<Peluquero>
+    fun findAllInRangeAtCoordinatesAndLikeNameOrWithTipos(rangoEnKm:Double,lat:Double,lng:Double, nombre: String, tipo: String, pageable: Pageable): Page<Peluquero>
 
 }
